+++ src/main/java/org/truffleruby/core/array/ObjectArrayStore.java
@@ -28,8 +28,8 @@ import org.truffleruby.core.array.ArrayStoreLibrary.ArrayAllocator;
 public class ObjectArrayStore {
 
     @ExportMessage
-    public static Object read(Object[] store, long index) {
-        return store[(int) index];
+    public static Object read(Object[] store, int index) {
+        return store[index];
     }
 
     @ExportMessage
@@ -58,18 +58,18 @@ public class ObjectArrayStore {
     }
 
     @ExportMessage
-    public static void write(Object[] store, long index, Object value) {
-        store[(int) index] = value;
+    public static void write(Object[] store, int index, Object value) {
+        store[index] = value;
     }
 
     @ExportMessage
-    public static long capacity(Object[] store) {
+    public static int capacity(Object[] store) {
         return store.length;
     }
 
     @ExportMessage
-    public static Object[] expand(Object[] store, long newCapacity) {
-        Object[] newStore = new Object[(int) newCapacity];
+    public static Object[] expand(Object[] store, int newCapacity) {
+        Object[] newStore = new Object[newCapacity];
         System.arraycopy(store, 0, newStore, 0, store.length);
         return newStore;
     }
@@ -78,35 +78,35 @@ public class ObjectArrayStore {
     static class CopyContents {
 
         @Specialization
-        static void copyContents(Object[] srcStore, long srcStart, Object[] destStore, long destStart, long length) {
-            System.arraycopy(srcStore, (int) srcStart, destStore, (int) destStart, (int) length);
+        static void copyContents(Object[] srcStore, int srcStart, Object[] destStore, int destStart, int length) {
+            System.arraycopy(srcStore, srcStart, destStore, destStart, length);
         }
 
         @Specialization
-        static void copyContents(Object[] srcStore, long srcStart, Object destStore, long destStart, long length,
+        static void copyContents(Object[] srcStore, int srcStart, Object destStore, int destStart, int length,
                 @CachedLibrary(limit = "5") ArrayStoreLibrary destStores) {
-            for (long i = srcStart; i < length; i++) {
-                destStores.write(destStore, destStart + i, srcStore[(int) (srcStart + i)]);
+            for (int i = srcStart; i < length; i++) {
+                destStores.write(destStore, destStart + i, srcStore[(srcStart + i)]);
             }
         }
     }
 
     @ExportMessage
-    public static Object[] copyStore(Object[] store, long length) {
-        return ArrayUtils.grow(store, (int) length);
+    public static Object[] copyStore(Object[] store, int length) {
+        return ArrayUtils.grow(store, length);
     }
 
     @ExportMessage
     @TruffleBoundary
-    public static void sort(Object[] store, long size) {
-        Arrays.sort(store, 0, (int) size);
+    public static void sort(Object[] store, int size) {
+        Arrays.sort(store, 0, size);
     }
 
     @ExportMessage
-    public static Iterable<Object> getIterable(Object[] store, long from, long length) {
+    public static Iterable<Object> getIterable(Object[] store, int from, int length) {
         return () -> new Iterator<Object>() {
 
-            private int n = (int) from;
+            private int n = from;
 
             @Override
             public boolean hasNext() {
@@ -177,8 +177,8 @@ public class ObjectArrayStore {
     private static class ObjectArrayAllocator extends ArrayAllocator {
 
         @Override
-        public Object[] allocate(long capacity) {
-            return new Object[(int) capacity];
+        public Object[] allocate(int capacity) {
+            return new Object[capacity];
         }
 
         @Override
